a little fix

+++ update.sh
@@ -1,6 +1,12 @@
-rm -rf vim vimrc
-cp -r ~/.vim .
-cp ~/.vimrc .
-mv .vim vim
-mv .vimrc vimrc
-rm -rf `find ./vim -name ".git*"`
+#!/bin/bash
+
+if [ $# -gt 0 ];then
+    dest=$1
+    src=.$1
+    rm -rf $dest
+    cp -r ~/$src .
+    mv $src $dest 
+    rm -rf `find ./vim -name ".git*"`
+else
+    echo 'Arg is wrong !'
+fi
